io/src/main/java/org/femtoframework/nio/ByteBufferAsOutputStream.java
package org.femtoframework.nio;

import java.io.IOException;
import java.io.OutputStream;
import java.nio.ByteBuffer;

/**
 * ByteBuffer到OutputStream
 *
 * @author fengyun
 * @version 1.00 2004-12-27 14:28:12
 */
public class ByteBufferAsOutputStream
    extends OutputStream
{
    private ByteBuffer buffer;

    /**
     * 构造
     *
     * @param buffer ByteBuffer
     */
    public ByteBufferAsOutputStream(ByteBuffer buffer)
    {
        if (buffer == null) {
            throw new IllegalArgumentException("Null byte buffer");
        }
        this.buffer = buffer;
    }

    /**
     * Writes the specified byte to this output stream. The general
     * contract for <code>write</code> is that one byte is written
     * to the output stream. The byte to be written is the eight
     * low-order bits of the argument <code>b</code>. The 24
     * high-order bits of <code>b</code> are ignored.
     * <p/>
     * Subclasses of <code>OutputStream</code> must provide an
     * implementation for this method.
     *
     * @param b the <code>byte</code>.
     * @throws IOException if an I/O error occurs. In particular,
     *                             an <code>IOException</code> may be thrown if the
     *                             output stream has been closed.
     */
    public void write(int b) throws IOException
    {
        buffer.put((byte)b);
    }

    /**
     * Writes <code>len</code> bytes from the specified byte array
     * starting at offset <code>off</code> to this output stream.
     * The general contract for <code>write(b, off, len)</code> is that
     * some of the bytes in the array <code>b</code> are written to the
     * output stream in order; element <code>b[off]</code> is the first
     * byte written and <code>b[off+len-1]</code> is the last byte written
     * by this operation.
     * <p/>
     * The <code>write</code> method of <code>OutputStream</code> calls
     * the write method of one argument on each of the bytes to be
     * written out. Subclasses are encouraged to override this method and
     * provide a more efficient implementation.
     * <p/>
     * If <code>b</code> is <code>null</code>, a
     * <code>NullPointerException</code> is thrown.
     * <p/>
     * If <code>off</code> is negative, or <code>len</code> is negative, or
     * <code>off+len</code> is greater than the length of the array
     * <code>b</code>, then an <tt>IndexOutOfBoundsException</tt> is thrown.
     *
     * @param b   the data.
     * @param off the start offset in the data.
     * @param len the number of bytes to write.
     * @throws IOException if an I/O error occurs. In particular,
     *                     an <code>IOException</code> is thrown if the output
     *                     stream is closed.
     */
    public void write(byte b[], int off, int len) throws IOException
    {
        int r = buffer.remaining();
        if (len > r) {
            throw new IOException("No space left:" + len + " remaining:" + r);
        }
        buffer.put(b, off, len);
    }
}
